requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile requirements.in
#
aniso8601==7.0.0          # via graphene
asgiref==3.2.10           # via django
certifi==2020.6.20        # via requests, sentry-sdk
chardet==3.0.4            # via requests
coreapi==2.3.3            # via -r requirements.in
coreschema==0.0.4         # via coreapi
django-anymail==7.2.1     # via -r requirements.in, django-ilmoitin
django-cors-headers==3.4.0  # via -r requirements.in
django-environ==0.4.5     # via -r requirements.in
django-filter==2.3.0      # via -r requirements.in
django-graphql-geojson==0.1.4  # via -r requirements.in
django-graphql-jwt==0.3.1  # via -r requirements.in
django-helusers==0.5.6    # via -r requirements.in
django-ilmoitin[graphql_api]==0.6.0  # via -r requirements.in
django-js-asset==1.2.2    # via django-mptt
django-mailer==2.0.1      # via -r requirements.in, django-ilmoitin
django-mptt==0.11.0       # via django-munigeo
git+https://github.com/a-rmz/django-munigeo.git@0.4.0-beta#egg=django-munigeo  # via -r requirements.in
django-parler-rest==2.1   # via django-munigeo
django-parler==2.1        # via -r requirements.in, django-ilmoitin, django-munigeo, django-parler-rest
django==3.1               # via -r requirements.in, django-anymail, django-cors-headers, django-filter, django-graphql-geojson, django-graphql-jwt, django-helusers, django-ilmoitin, django-mailer, django-mptt, django-munigeo, djangorestframework, drf-oidc-auth, graphene-django
djangorestframework==3.11.1  # via django-parler-rest, drf-oidc-auth
drf-oidc-auth==0.10.0     # via django-helusers
ecdsa==0.14.1             # via python-jose
factory-boy==3.0.1        # via -r requirements.in
faker==4.1.1              # via factory-boy
future==0.18.2            # via pyjwkest
graphene-django==2.13.0   # via -r requirements.in, django-graphql-geojson, django-graphql-jwt, django-ilmoitin
graphene-federation==0.1.0  # via -r requirements.in
graphene-file-upload==1.2.2  # via -r requirements.in
graphene==2.1.8           # via django-ilmoitin, graphene-django, graphene-federation
graphql-core==2.3.2       # via django-graphql-jwt, django-ilmoitin, graphene, graphene-django, graphql-relay
graphql-relay==2.0.1      # via django-ilmoitin, graphene
idna==2.10                # via requests
itypes==1.2.0             # via coreapi
jinja2==2.11.2            # via coreschema, django-ilmoitin
lockfile==0.12.2          # via django-mailer
markupsafe==1.1.1         # via jinja2
pillow==7.2.0             # via -r requirements.in
promise==2.3              # via graphene-django, graphql-core, graphql-relay
psycopg2-binary==2.8.5    # via -r requirements.in
pyasn1==0.4.8             # via python-jose, rsa
pycryptodomex==3.9.8      # via pyjwkest
pyjwkest==1.4.2           # via drf-oidc-auth
pyjwt==1.7.1              # via django-graphql-jwt
python-dateutil==2.8.1    # via faker
python-jose==3.2.0        # via django-helusers
pytz==2020.1              # via django
pyyaml==5.3.1             # via django-munigeo
requests-cache==0.5.2     # via django-munigeo
requests==2.24.0          # via coreapi, django-anymail, django-helusers, django-munigeo, pyjwkest, requests-cache
rsa==4.6                  # via python-jose
rx==1.6.1                 # via graphql-core
sentry-sdk==0.16.5        # via -r requirements.in
singledispatch==3.4.0.3   # via graphene-django
six==1.15.0               # via django-anymail, django-mailer, django-munigeo, django-parler, ecdsa, graphene, graphene-django, graphene-file-upload, graphql-core, graphql-relay, promise, pyjwkest, python-dateutil, python-jose, singledispatch
sqlparse==0.3.1           # via django
text-unidecode==1.3       # via faker
unidecode==1.1.1          # via graphene-django
uritemplate==3.0.1        # via coreapi
urllib3==1.25.10          # via requests, sentry-sdk
xlsxwriter==1.3.3         # via -r requirements.in
